+++ Source/QiscusCore/Database/RoomStorage.swift
@@ -254,7 +254,7 @@ extension RoomStorage {
         }
         result.id            = core.id
         result.uniqueId      = core.uniqueId
-        result.unreadCount   = Int16(core.unreadCount)
+        result.unreadCount   = core.unreadCount
         result.name          = core.name
         result.avatarUrl     = core.avatarUrl?.absoluteString ?? ""
         result.options       = core.options
